fix(torghut-ta): parse fractional crypto envelopes end-to-end (#3394)
* fix(torghut): preserve knative creator annotation for gitops sync

* fix(torghut-ta): parse fractional crypto envelopes end-to-end

+++ services/dorvud/technical-analysis/src/main/kotlin/ai/proompteng/dorvud/ta/stream/MicroBarAggregator.kt
@@ -105,7 +105,7 @@ class MicroBarAggregator {
     high = trade.p,
     low = trade.p,
     close = trade.p,
-    volume = trade.s.toDouble(),
+    volume = trade.s,
     vwapNumerator = trade.p * trade.s,
     count = 1,
   )
